requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2024.8.30
click==8.1.7
elastic-transport==8.15.1
elasticsearch==8.16.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
setuptools==75.6.0
urllib3==2.2.3
Werkzeug==3.0.1
wheel==0.41.3
